graphql-gateway/requirements.txt
fastapi
uvicorn
strawberry-graphql[fastapi]
httpx
python-dotenv